azure-mfa-multiagent/backend_flask/requirements.txt
Flask==2.2.5
requests==2.31.0
python-dotenv==1.0.0
bcrypt==4.0.1
pyjwt==2.8.0
